pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
mysqlclient = "^2.1.1"
dj-database-url = "^1.2.0"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
